kaito GA changes
