+++ src/commands/aksKaito/aksKaitoCreateCRD.ts
@@ -6,9 +6,9 @@ import { getKubernetesClusterInfo } from "../utils/clusters";
 import { getReadySessionProvider } from "../../auth/azureAuth";
 import { KaitoModelsPanelDataProvider } from "../../panels/KaitoModelsPanel";
 import { KaitoModelsPanel } from "../../panels/KaitoModelsPanel";
-import { filterPodName, getAksClusterTreeNode } from "../utils/clusters";
+import { getAksClusterTreeNode } from "../utils/clusters";
 import { failed } from "../utils/errorable";
-import { getExtension, longRunning } from "../utils/host";
+import { getExtension } from "../utils/host";
 
 export default async function aksKaitoCreateCRD(_context: IActionContext, target: unknown): Promise<void> {
     const cloudExplorer = await k8s.extension.cloudExplorer.v1;
@@ -48,23 +48,6 @@ export default async function aksKaitoCreateCRD(_context: IActionContext, target
     }
 
     const { name: clusterName, armId, subscriptionId, resourceGroupName } = clusterNode.result;
-
-    // "kaito-" is assuming kaito pods are still named kaito-workspace & kaito-gpu-provisioner
-    const filterKaitoPodNames = await longRunning(`Checking if KAITO is installed.`, () => {
-        return filterPodName(sessionProvider.result, kubectl, subscriptionId, resourceGroupName, clusterName, "kaito-");
-    });
-
-    if (failed(filterKaitoPodNames)) {
-        vscode.window.showErrorMessage(filterKaitoPodNames.error);
-        return;
-    }
-
-    if (filterKaitoPodNames.result.length === 0) {
-        vscode.window.showWarningMessage(
-            `Please install KAITO for cluster ${clusterName}. \n \n KAITO Workspace generation is only enabled when KAITO is installed.`,
-        );
-        return;
-    }
 
     const clusterInfo = await getKubernetesClusterInfo(sessionProvider.result, target, cloudExplorer, clusterExplorer);
     if (failed(clusterInfo)) {
